Reworked game() & playRound()
I chnaged the parameter of my playRound() to incude the prompt'Rock, Paper
or Scissors?'
I also included the prompt in the variable playerChoice and defined it
together with  computerChoice variable in the game() function.
This fixed an issue i was having; when calling the game() function more
than once it was giivng me the same computerChoice and using the same
playerChoice from the first call.